requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
google-cloud-storage==2.13.0
openai-whisper==20231117
ここまではダウンロード済み